# shared/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FINN Licensing API
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  message: string;
}

export interface MessageResponse {
  message: string;
}

export interface AuthUser {
  id: string;
  email: string;
  name: string;
}

export interface DashboardStats {
  clientCount: number;
  productCount: number;
  licenseCount: number;
  activeLicenseCount: number;
}

export interface Client {
  id: number;
  name: string;
  company?: string | null;
  email?: string | null;
  notes?: string | null;
  licenseCount: number;
  createdAt: string;
}

export interface ClientInput {
  name: string;
  company?: string | null;
  email?: string | null;
  notes?: string | null;
}

export interface Product {
  id: number;
  name: string;
  slug: string;
  githubRepo: string;
  description?: string | null;
  latestVersion?: string | null;
  releaseDate?: string | null;
  changelog?: string | null;
  downloadUrl?: string | null;
  requiresWp?: string | null;
  testedWp?: string | null;
  requiresPhp?: string | null;
  lastChecked?: string | null;
  createdAt: string;
}

export interface ProductInput {
  name: string;
  slug: string;
  githubRepo: string;
  description?: string | null;
  requiresWp?: string | null;
  testedWp?: string | null;
  requiresPhp?: string | null;
}

export interface PollResult {
  success: boolean;
  message: string;
  product?: Product;
}

export type LicensePluginAccess =
  (typeof LicensePluginAccess)[keyof typeof LicensePluginAccess];

export const LicensePluginAccess = {
  all: "all",
  specific: "specific",
} as const;

export type LicenseStatus = (typeof LicenseStatus)[keyof typeof LicenseStatus];

export const LicenseStatus = {
  active: "active",
  revoked: "revoked",
} as const;

export interface License {
  id: number;
  licenseKeyPreview: string;
  clientId?: number | null;
  clientName?: string | null;
  domain: string;
  pluginAccess: LicensePluginAccess;
  productIds?: number[] | null;
  status: LicenseStatus;
  createdAt: string;
}

export type LicenseInputPluginAccess =
  (typeof LicenseInputPluginAccess)[keyof typeof LicenseInputPluginAccess];

export const LicenseInputPluginAccess = {
  all: "all",
  specific: "specific",
} as const;

export type LicenseInputStatus =
  (typeof LicenseInputStatus)[keyof typeof LicenseInputStatus];

export const LicenseInputStatus = {
  active: "active",
  revoked: "revoked",
} as const;

export interface LicenseInput {
  clientId: number;
  domain: string;
  pluginAccess: LicenseInputPluginAccess;
  productIds?: number[] | null;
  status?: LicenseInputStatus;
}

export interface LicenseCreated {
  license: License;
  licenseKey: string;
}

export interface PublicStatus {
  status: string;
  version: string;
}

export interface ValidateRequest {
  key: string;
  fingerprint: string;
}

export type ValidateResponseData = {
  valid: boolean;
};

export interface ValidateResponse {
  data: ValidateResponseData;
}

export interface PublicProduct {
  id: string;
  name: string;
  slug: string;
  version?: string | null;
  description?: string | null;
}

export interface Release {
  id: number;
  productId: number;
  version: string;
  tagName: string;
  changelog?: string | null;
  downloadUrl?: string | null;
  zipballUrl?: string | null;
  publishedAt?: string | null;
  createdAt: string;
}

export type UpdateCheckResponseSections = {
  changelog?: string;
} | null;

export interface UpdateCheckResponse {
  version?: string | null;
  download_url?: string | null;
  slug?: string | null;
  tested?: string | null;
  requires?: string | null;
  requires_php?: string | null;
  sections?: UpdateCheckResponseSections;
}

export type PublicProducts200 = {
  data: PublicProduct[];
};

export type PublicUpdateCheckParams = {
  product_id: string;
  license: string;
  fingerprint: string;
};

export type PublicDownloadParams = {
  product_id: string;
  license: string;
  fingerprint: string;
};