# shared/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: FINN Licensing API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/products/{id}/releases:
    get:
      operationId: listProductReleases
      tags: [products]
      summary: List all releases for a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Release list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Release"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/products/{id}/poll:
    post:
      operationId: pollProduct
      tags: [products]
      summary: Poll GitHub for latest release
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Poll result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PollResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        githubRepo:
          type: string
        description:
          type: string
          nullable: true
        latestVersion:
          type: string
          nullable: true
        releaseDate:
          type: string
          format: date-time
          nullable: true
        changelog:
          type: string
          nullable: true
        downloadUrl:
          type: string
          nullable: true
        requiresWp:
          type: string
          nullable: true
        testedWp:
          type: string
          nullable: true
        requiresPhp:
          type: string
          nullable: true
        lastChecked:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - slug
        - githubRepo
        - createdAt
    ProductInput:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        githubRepo:
          type: string
        description:
          type: string
          nullable: true
        requiresWp:
          type: string
          nullable: true
        testedWp:
          type: string
          nullable: true
        requiresPhp:
          type: string
          nullable: true
      required:
        - name
        - slug
        - githubRepo
    PollResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        product:
          $ref: "#/components/schemas/Product"
      required:
        - success
        - message
...
    Release:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        version:
          type: string
        tagName:
          type: string
        changelog:
          type: string
          nullable: true
        downloadUrl:
          type: string
          nullable: true
        zipballUrl:
          type: string
          nullable: true
        publishedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - productId
        - version
        - tagName
        - createdAt
...